idk man, you tell me

+++ mkiso.sh
@@ -8,11 +8,15 @@ if [ "$(uname -s)" != 'Linux' ]; then
     exit 1
 fi
 
-time nom build .
+time nom build --show-trace .
 
-for resultISO in $(basename result/iso/nixos-*-"$(uname -m)"-linux.iso); do
-    cp result/iso/"${resultISO}" .
-    sha512sum "${resultISO}" | awk '{print $1}' | tee "${resultISO}.sha512"
+for resultISO in $(basename result/iso/nixos-*-linux.iso); do
+    if [ ! -f "${resultISO}" ]; then
+        cp result/iso/"${resultISO}" .
+    fi
+    if [ ! -f "${resultISO}.sha512" ]; then
+        sha512sum "${resultISO}" | awk '{print $1}' 1> "${resultISO}.sha512"
+    fi
     chown "${USER}:${USER}" "${resultISO}"
     chmod 644 "${resultISO}"
 done
